+++ packages/dapp/.eslintrc.cjs
@@ -1,9 +1,0 @@
-module.exports = {
-  root: true,
-  extends: ['next/core-web-vitals', '../../.eslintrc.cjs'],
-  parserOptions: {
-    tsconfigRootDir: __dirname,
-    project: './tsconfig.eslint.json',
-    sourceType: 'module',
-  },
-};