+++ packages/dapp/global.d.ts
@@ -1,0 +1,7 @@
+declare global {
+  interface Window {
+    ethereum: import('ethers').providers.ExternalProvider;
+  }
+}
+
+export {};