+++ packages/dapp/next.config.js
@@ -1,4 +1,4 @@
-import StylelintPlugin from 'stylelint-webpack-plugin';
+const StylelintPlugin = require('stylelint-webpack-plugin');
 
 // Content-Security-Policy
 // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy
@@ -17,14 +17,16 @@ const isProd = process.env.NODE_ENV === 'production';
  * @type {import('next').NextConfig}
  */
 const nextConfig = {
-  // FIXME: On release change to /ssi-snap
-  basePath: '/dapp',
+  // FIXME: On release change to /masca
+  basePath: '/ssi-snap',
   reactStrictMode: true,
   swcMinify: true,
   // https://nextjs.org/docs/advanced-features/output-file-tracing#automatically-copying-traced-files
   output: 'standalone',
   // https://nextjs.org/docs/messages/next-image-unconfigured-host
   images: {
+    // Disable image optimization
+    unoptimized: true,
     domains: [],
   },
   experimental: {
@@ -32,6 +34,7 @@ const nextConfig = {
       { loader: '@next/font/google', options: { subsets: ['latin'] } },
     ],
   },
+
   // Security headers and CSP
   // https://nextjs.org/docs/advanced-features/security-headers
   // Not supported with `next export`
@@ -85,4 +88,4 @@ const nextConfig = {
   },
 };
 
-export default nextConfig;
+module.exports = nextConfig;