+++ packages/snap/src/rpc/vc/saveVC.ts
@@ -1,5 +1,8 @@
-import { SaveVCRequestParams } from '@blockchain-lab-um/ssi-snap-types';
-import { IDataManagerSaveResult } from '@blockchain-lab-um/veramo-vc-manager';
+import {
+  SaveVCRequestParams,
+  SaveVCRequestResult,
+} from '@blockchain-lab-um/ssi-snap-types';
+import { copyable, divider, heading, panel, text } from '@metamask/snaps-ui';
 
 import { ApiParams } from '../../interfaces';
 import { snapConfirm } from '../../utils/snapUtils';
@@ -8,18 +11,20 @@ import { veramoSaveVC } from '../../utils/veramoUtils';
 export async function saveVC(
   params: ApiParams,
   { verifiableCredential, options }: SaveVCRequestParams
-): Promise<IDataManagerSaveResult[]> {
+): Promise<SaveVCRequestResult[]> {
   const { store = 'snap' } = options || {};
   const { snap, ethereum } = params;
-  const promptObj = {
-    prompt: 'Save VC',
-    description: `Would you like to save the following VC in ${
-      typeof store === 'string' ? store : store.join(', ')
-    }?`,
-    textAreaContent: JSON.stringify(verifiableCredential).substring(0, 100),
-  };
 
-  if (snapConfirm(snap, promptObj)) {
+  const content = panel([
+    heading('Save VC'),
+    text('Would you like to save following VC?'),
+    divider(),
+    text(`Store(s): ${typeof store === 'string' ? store : store.join(', ')}`),
+    text(`VC:`),
+    copyable(JSON.stringify(verifiableCredential, null, 2)),
+  ]);
+
+  if (await snapConfirm(snap, content)) {
     const res = await veramoSaveVC({
       snap,
       ethereum,