+++ packages/snap/src/rpc/vcStore/setVCStore.ts
@@ -1,4 +1,5 @@
 import { SetVCStoreRequestParams } from '@blockchain-lab-um/ssi-snap-types';
+import { heading, panel, text } from '@metamask/snaps-ui';
 
 import { ApiParams } from '../../interfaces';
 import { snapConfirm } from '../../utils/snapUtils';
@@ -10,14 +11,11 @@ export async function setVCStore(
 ): Promise<boolean> {
   const { state, snap, account } = params;
   if (store !== 'snap') {
-    const promptObj = {
-      prompt: 'Change vcStore plugin',
-      description: `Would you like to ${
-        value ? 'enable' : 'disable'
-      } ${store} vcStore plugin?`,
-      textAreaContent: `Content`,
-    };
-    if (snapConfirm(snap, promptObj)) {
+    const content = panel([
+      heading('Manage VCStore Plugin'),
+      text(`Would you like to ${value ? 'enable' : 'disable'} ${store}?`),
+    ]);
+    if (await snapConfirm(snap, content)) {
       state.accountState[account].accountConfig.ssi.vcStore[store] = value;
       await updateSnapState(snap, state);
       return true;