+++ packages/snap/src/utils/veramoUtils.ts
@@ -3,19 +3,19 @@ import {
   CreateVPRequestParams,
   QueryVCsOptions,
   QueryVCsRequestResult,
+  SaveVCRequestResult,
   VerifyDataRequestParams,
 } from '@blockchain-lab-um/ssi-snap-types';
-import {
-  Filter,
-  IDataManagerSaveResult,
-} from '@blockchain-lab-um/veramo-vc-manager';
+import { Filter } from '@blockchain-lab-um/veramo-vc-manager';
 import { BIP44CoinTypeNode } from '@metamask/key-tree';
 import { MetaMaskInpageProvider } from '@metamask/providers';
 import { SnapsGlobalObject } from '@metamask/snaps-types';
+import { copyable, divider, heading, panel, text } from '@metamask/snaps-ui';
 import {
   IIdentifier,
   IVerifyResult,
   MinimalImportableKey,
+  VerifiableCredential,
   VerifiablePresentation,
   W3CVerifiableCredential,
 } from '@veramo/core';
@@ -31,14 +31,33 @@ export async function veramoSaveVC(args: {
   ethereum: MetaMaskInpageProvider;
   verifiableCredential: W3CVerifiableCredential;
   store: AvailableVCStores | AvailableVCStores[];
-}): Promise<IDataManagerSaveResult[]> {
+}): Promise<SaveVCRequestResult[]> {
   const { snap, ethereum, store, verifiableCredential } = args;
   const agent = await getAgent(snap, ethereum);
-  const res = await agent.save({
+  const result = await agent.save({
     data: verifiableCredential,
     options: { store },
   });
-  return res;
+
+  const vcs = new Map<string, SaveVCRequestResult>();
+
+  for (const vc of result) {
+    if (!vc.store) {
+      throw new Error('Missing store in VC metadata');
+    }
+
+    const existingVC = vcs.get(vc.id);
+    if (existingVC) {
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-call
+      existingVC.store?.push(vc.store);
+    } else {
+      vcs.set(vc.id, {
+        id: vc.id,
+        store: [vc.store],
+      });
+    }
+  }
+  return [...vcs.values()];
 }
 
 export const veramoImportMetaMaskAccount = async (
@@ -138,7 +157,7 @@ export async function veramoQueryVCs(args: {
       }
     } else {
       vcs.set(vc.metadata.id, {
-        data: vc.data as W3CVerifiableCredential,
+        data: vc.data as VerifiableCredential,
         metadata: {
           id: vc.metadata.id,
           ...(options.returnStore && { store: [vc.metadata.store as string] }),
@@ -152,7 +171,7 @@ export async function veramoQueryVCs(args: {
 export async function veramoCreateVP(
   params: ApiParams,
   createVPParams: CreateVPRequestParams
-): Promise<VerifiablePresentation | null> {
+): Promise<VerifiablePresentation> {
   const vcsMetadata = createVPParams.vcs;
   const domain = createVPParams.proofOptions?.domain;
   const challenge = createVPParams.proofOptions?.challenge;
@@ -184,14 +203,20 @@ export async function veramoCreateVP(
     }
   }
 
-  if (vcs.length === 0) return null;
+  if (vcs.length === 0) {
+    throw new Error('VC does not exist');
+  }
+
   const config = state.snapConfig;
-  const promptObj = {
-    prompt: 'Alert',
-    description: 'Do you wish to create a VP from the following VC?',
-    textAreaContent: 'Multiple VCs',
-  };
-  if (config.dApp.disablePopups || snapConfirm(snap, promptObj)) {
+  const content = panel([
+    heading('Create VP'),
+    text('Would you like to create a VP from the following VC(s)?'),
+    divider(),
+    text(`VC(s):`),
+    ...vcs.map((vc) => copyable(JSON.stringify(vc, null, 2))),
+  ]);
+
+  if (config.dApp.disablePopups || (await snapConfirm(snap, content))) {
     const vp = await agent.createVerifiablePresentation({
       presentation: {
         holder: identifier.did,
@@ -204,7 +229,8 @@ export async function veramoCreateVP(
     });
     return vp;
   }
-  return null;
+
+  throw new Error('User rejected create VP request');
 }
 
 export async function veramoVerifyData(args: {