+++ packages/snap/tests/testUtils/constants.ts
@@ -1,4 +1,5 @@
 import { JsonBIP44CoinTypeNode } from '@metamask/key-tree';
+import { heading, panel, text } from '@metamask/snaps-ui';
 import {
   DIDDocument,
   DIDResolutionResult,
@@ -488,3 +489,8 @@ export const derivedKeyChainCode =
   '0xbbfe64aaa2157420865e979f19bdd0da7af456d075d7e4d4bb1ee92e548ac440';
 export const derivedKeyDerivationPath =
   "m / bip32:44' / bip32:60' / bip32:0' / bip32:0";
+
+export const content = panel([
+  heading('Title of the panel'),
+  text('Text of the panel'),
+]);