+++ scripts/build-docker.sh
@@ -2,7 +2,10 @@
 
 projects="\
 @blockchain-lab-um/ssi-snap-docs,\
-@blockchain-lab-um/website,\
+@blockchain-lab-um/dapp,\
+@blockchain-lab-um/utils,\
+@blockchain-lab-um/ssi-snap-types,\
+@blockchain-lab-um/ssi-snap-connector,\
 @blockchain-lab-um/oidc-issuer,\
 @blockchain-lab-um/oidc-verifier,\
 @blockchain-lab-um/oidc-rp-plugin\